fix: moved static class to isAvailable method
