+++ src/Resources/config/services/express_checkout.xml
@@ -13,6 +13,7 @@
             <argument type="service" id="Shopware\Core\System\SystemConfig\SystemConfigService"/>
             <argument type="service" id="Swag\PayPal\Setting\Service\CredentialsUtil"/>
             <argument type="service" id="Swag\PayPal\Checkout\Cart\Service\CartPriceService"/>
+            <argument type="service" id="Swag\PayPal\Util\Lifecycle\Method\PayLaterMethodData"/>
         </service>
 
         <service id="Swag\PayPal\Checkout\ExpressCheckout\SalesChannel\ExpressPrepareCheckoutRoute" public="true">
