+++ src/Resources/config/services/installment.xml
@@ -11,6 +11,7 @@
             <argument type="service" id="Swag\PayPal\Installment\Banner\Service\BannerDataService"/>
             <argument type="service" id="Swag\PayPal\Checkout\Cart\Service\ExcludedProductValidator"/>
             <argument type="service" id="monolog.logger.paypal"/>
+            <argument type="service" id="Swag\PayPal\Util\Lifecycle\Method\PayLaterMethodData"/>
             <tag name="kernel.event_subscriber"/>
         </service>
